Allow per-page transitions for presentations.
Use a conditional master-detail interface in the options sheet, with a table displaying the pages using the thumbnails from the main window, controlled through bindings. The array of transitions are saved in the EAs. Make changes to the transitions undoable, both locally in the transition sheet and for the document. 

git-svn-id: https://svn.code.sf.net/p/skim-app/code/trunk@5529 0fceea05-a30d-0410-8a8b-80ef821fb0a1